+++ R/to_postgres.R
@@ -14,7 +14,8 @@
 #' @param table_name string Name of new table in Postgres, required.
 #' @param schema_name string Name of final schema if not public, default is public.
 #' @param primary_keys string Fields to for primary keys, used in a SQL statement, default is NA.
-#' @return Paste copy of output here.  You will get a generic message from RPostgres, check
+#'
+#' @return You will get a generic message from RPostgres, check
 #' Postgres server for the data to ensure that the data was transferred.
 #'
 #' @examples
@@ -24,6 +25,9 @@
 #' }
 #'
 #' @import rlang
+#' @importFrom RPostgres dbExistsTable dbRemoveTable dbWriteTable
+#' @importFrom DBI Id
+#' @importFrom stringr str_to_lower
 #' @export
 
 # RPostgres::dbWriteTable(con, name = DBI::Id(schema = "landing", table = "iris"), value = iris, overwrite = TRUE)